+++ jscomp/gentype_tests/typescript-react-example/src/FirstClassModulesInterface.gen.tsx
@@ -3,13 +3,6 @@
 /* eslint-disable */
 /* tslint:disable */
 
-import * as FirstClassModulesInterfaceBS__Es6Import from './FirstClassModulesInterface.bs';
-const FirstClassModulesInterfaceBS: any = FirstClassModulesInterfaceBS__Es6Import;
-
 export type record = { readonly x: number; readonly y: string };
 
 export type firstClassModule = { readonly x: number };
-
-export const MT_x: number = FirstClassModulesInterfaceBS.MT.x;
-
-export const MT: { x: number } = FirstClassModulesInterfaceBS.MT