+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 
 - Fix accidental removal of `Belt.Result.Ok` and `Belt.Result.Error` constructors in rc.5 https://github.com/rescript-lang/rescript-compiler/pull/6514
 - Add missing check that the runtime representation of variants matches implementation and interface. https://github.com/rescript-lang/rescript-compiler/pull/6513/files
+- GenType: only export types (not values) from module types. https://github.com/rescript-lang/rescript-compiler/pull/6516
 
 # 11.0.0-rc.7